complete metodos, pero hay error en createCollection clse BookCollectionManager

+++ src/edu/uaslp/examen/Main.java
@@ -37,7 +37,7 @@ public class Main {
         book.setTitle(title);
         book.setIsbn(isbn);
         book.setAuthor(author);
-        book.setStars(stars);
+        book.setStars((short) stars);
 
         return book;
     }
